Comments removed, code cleaning

+++ SourceCode/GPS/Properties/Settings.cs
@@ -275,7 +275,7 @@ namespace AgOpenGPS.Properties
         public bool setDisplay_isLineSmooth = false;
         public Size setWindow_tramLineSize = new Size(921, 676);
         public bool setAutoSwitchDualFixOn = false;
-        public int setAutoSwitchDualFixSpeed = 2;
+        public double setAutoSwitchDualFixSpeed = 2.0;
 
         public LoadResult Load()
         {
